requirements.txt
bcrypt==4.2.0
blinker==1.8.2
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
Flask[async]==3.0.3
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pycparser==2.22
PyJWT==2.9.0
pyodbc==5.1.0
python-decouple==3.8
pytz==2024.2
Werkzeug==3.0.4
requests==2.32.3
aiohttp==3.9.5
translate==3.6.1
cryptography==43.0.1